refactor(ui): rehome typography primitives
