+++ npm/ui/core/scripts/check-size.mjs
@@ -19,7 +19,7 @@ const budgets = new Map([
   ["tabs.mjs", 15_850],
   ["stepper.mjs", 17_600],
   ["card.mjs", 900],
-  ["code.mjs", 850],
+  ["code.mjs", 875],
   ["cluster.mjs", 1_550],
   ["container.mjs", 1_450],
   ["grid.mjs", 1_650],
@@ -54,7 +54,7 @@ const budgets = new Map([
   ["meter.mjs", 3_700],
   ["native-select.mjs", 4_800],
   ["heading.mjs", 1_000],
-  ["kbd.mjs", 850],
+  ["kbd.mjs", 875],
   ["list.mjs", 900],
   ["listbox.mjs", 16_000],
   ["pagination.mjs", 6_750],
