+++ npm/ui/core/src/directory-layout.test.ts
@@ -17,7 +17,6 @@ const grandfatheredRootSfcFiles = [
   "avatar.vue",
   "badge.vue",
   "block-ui.vue",
-  "blockquote.vue",
   "breadcrumb-item.vue",
   "breadcrumb-link.vue",
   "breadcrumb-list.vue",
@@ -26,7 +25,6 @@ const grandfatheredRootSfcFiles = [
   "card.vue",
   "checkbox-control.vue",
   "cluster.vue",
-  "code.vue",
   "collapsible-content.vue",
   "collapsible-root.vue",
   "collapsible-trigger.vue",
@@ -39,8 +37,6 @@ const grandfatheredRootSfcFiles = [
   "field-label.vue",
   "field.vue",
   "grid.vue",
-  "heading.vue",
-  "kbd.vue",
   "link-anchor.vue",
   "list.vue",
   "listbox-item.vue",
@@ -79,7 +75,6 @@ const grandfatheredRootSfcFiles = [
   "tabs-root.vue",
   "tabs-trigger.vue",
   "text-input.vue",
-  "text.vue",
   "textarea-control.vue",
   "toggle-button.vue",
   "toggle-group-item.vue",
