+++ npm/ui/core/src/families/typography/code/code.vue
@@ -3,7 +3,7 @@
     :is="as"
     ref="element"
     part="root"
-    data-vize-ui="kbd"
+    data-vize-ui="code"
     :data-size="sizeState"
     :data-variant="variantState"
     :data-tone="toneState"
@@ -16,66 +16,66 @@
 import { computed, useTemplateRef } from "vue";
 
 import type {
-  KbdElement,
-  KbdExpose,
-  KbdSize,
-  KbdSlotState,
-  KbdTone,
-  KbdVariant,
-} from "./kbd-types.ts";
-import type { PrimitiveAs } from "./primitive.ts";
+  CodeElement,
+  CodeExpose,
+  CodeSize,
+  CodeSlotState,
+  CodeTone,
+  CodeVariant,
+} from "./code-types.ts";
+import type { PrimitiveAs } from "../../../primitive.ts";
 
 const {
-  as = "kbd",
+  as = "code",
   size = "md",
-  variant = "key",
+  variant = "inline",
   tone = "neutral",
 } = defineProps<{
   /**
    * Native element, custom element, or component to render.
    *
-   * @default "kbd"
+   * @default "code"
    */
   readonly as?: PrimitiveAs;
 
   /**
-   * Consumer visual-size token mirrored to `data-size`; no CSS is emitted.
+   * Consumer code-size token mirrored to `data-size`; no CSS is emitted.
    *
    * @default "md"
    */
-  readonly size?: KbdSize;
+  readonly size?: CodeSize;
 
   /**
-   * Keyboard presentation token mirrored to `data-variant`; no CSS is emitted.
+   * Code presentation token mirrored to `data-variant`; no CSS is emitted.
    *
-   * @default "key"
+   * @default "inline"
    */
-  readonly variant?: KbdVariant;
+  readonly variant?: CodeVariant;
 
   /**
    * Consumer tone token mirrored to `data-tone`; no CSS is emitted.
    *
    * @default "neutral"
    */
-  readonly tone?: KbdTone;
+  readonly tone?: CodeTone;
 }>();
 
 defineSlots<{
-  /** Renders keyboard input content with current presentation hook state. */
-  default(props: KbdSlotState): unknown;
+  /** Renders code content with current typography hook state. */
+  default(props: CodeSlotState): unknown;
 }>();
 
-const element = useTemplateRef<KbdElement>("element");
+const element = useTemplateRef<CodeElement>("element");
 const sizeState = computed(() => size);
 const variantState = computed(() => variant);
 const toneState = computed(() => tone);
-const slotState = computed<KbdSlotState>(() => ({
+const slotState = computed<CodeSlotState>(() => ({
   size: sizeState.value,
   tone: toneState.value,
   variant: variantState.value,
 }));
 
-type KbdSetupExpose = Omit<KbdExpose, "element" | "size" | "tone" | "variant"> & {
+type CodeSetupExpose = Omit<CodeExpose, "element" | "size" | "tone" | "variant"> & {
   readonly element: typeof element;
   readonly size: typeof sizeState;
   readonly tone: typeof toneState;
@@ -87,11 +87,11 @@ const exposed = {
   size: sizeState,
   tone: toneState,
   variant: variantState,
-} satisfies KbdSetupExpose;
+} satisfies CodeSetupExpose;
 
 defineExpose(exposed);
 </script>
 
 <style scoped>
-/* Headless by design. Keycap styling, spacing, and separators remain consumer-owned. */
+/* Headless by design. Code font, wrapping, syntax color, and block layout remain consumer-owned. */
 </style>
