+++ npm/ui/core/src/families/typography/heading/heading.vue
@@ -1,9 +1,10 @@
 <template>
   <component
-    :is="as"
+    :is="resolvedAs"
     ref="element"
     part="root"
-    data-vize-ui="text"
+    data-vize-ui="heading"
+    :data-level="levelState"
     :data-size="sizeState"
     :data-weight="weightState"
     :data-tone="toneState"
@@ -16,50 +17,60 @@
 <script setup lang="ts">
 import { computed, useTemplateRef } from "vue";
 
-import type { PrimitiveAs } from "./primitive.ts";
+import type { PrimitiveAs } from "../../../primitive.ts";
 import type {
-  TextElement,
-  TextExpose,
-  TextSlotState,
-  TextSize,
-  TextTone,
-  TextWeight,
-} from "./text-types.ts";
+  HeadingElement,
+  HeadingExpose,
+  HeadingLevel,
+  HeadingSize,
+  HeadingSlotState,
+  HeadingTone,
+  HeadingWeight,
+} from "./heading-types.ts";
 
 const {
-  as = "span",
+  as,
+  level = 2,
   size = "md",
-  weight = "regular",
+  weight = "semibold",
   tone = "neutral",
   truncate = false,
 } = defineProps<{
   /**
-   * Native element, custom element, or component to render.
+   * Native element, custom element, or component to render. When omitted,
+   * Heading renders the native `h${level}` element.
    *
-   * @default "span"
+   * @default undefined
    */
   readonly as?: PrimitiveAs;
 
   /**
-   * Consumer text-size token mirrored to `data-size`; no CSS is emitted.
+   * Semantic heading level used for the default native host and `data-level`.
+   *
+   * @default 2
+   */
+  readonly level?: HeadingLevel;
+
+  /**
+   * Consumer visual-size token mirrored to `data-size`; no CSS is emitted.
    *
    * @default "md"
    */
-  readonly size?: TextSize;
+  readonly size?: HeadingSize;
 
   /**
    * Consumer font-weight token mirrored to `data-weight`; no CSS is emitted.
    *
-   * @default "regular"
+   * @default "semibold"
    */
-  readonly weight?: TextWeight;
+  readonly weight?: HeadingWeight;
 
   /**
    * Consumer tone token mirrored to `data-tone`; no CSS is emitted.
    *
    * @default "neutral"
    */
-  readonly tone?: TextTone;
+  readonly tone?: HeadingTone;
 
   /**
    * Consumer truncation hook mirrored to `data-truncate`; no CSS is emitted.
@@ -70,24 +81,31 @@ const {
 }>();
 
 defineSlots<{
-  /** Renders text content with current typography hook state. */
-  default(props: TextSlotState): unknown;
+  /** Renders heading content with current semantic and typography hook state. */
+  default(props: HeadingSlotState): unknown;
 }>();
 
-const element = useTemplateRef<TextElement>("element");
+const element = useTemplateRef<HeadingElement>("element");
+const levelState = computed(() => level);
 const sizeState = computed(() => size);
 const weightState = computed(() => weight);
 const toneState = computed(() => tone);
 const truncateState = computed(() => truncate);
-const slotState = computed<TextSlotState>(() => ({
+const resolvedAs = computed(() => as ?? `h${levelState.value}`);
+const slotState = computed<HeadingSlotState>(() => ({
+  level: levelState.value,
   size: sizeState.value,
   tone: toneState.value,
   truncate: truncateState.value,
   weight: weightState.value,
 }));
 
-type TextSetupExpose = Omit<TextExpose, "element" | "size" | "tone" | "truncate" | "weight"> & {
+type HeadingSetupExpose = Omit<
+  HeadingExpose,
+  "element" | "level" | "size" | "tone" | "truncate" | "weight"
+> & {
   readonly element: typeof element;
+  readonly level: typeof levelState;
   readonly size: typeof sizeState;
   readonly tone: typeof toneState;
   readonly truncate: typeof truncateState;
@@ -96,15 +114,16 @@ type TextSetupExpose = Omit<TextExpose, "element" | "size" | "tone" | "truncate"
 
 const exposed = {
   element,
+  level: levelState,
   size: sizeState,
   tone: toneState,
   truncate: truncateState,
   weight: weightState,
-} satisfies TextSetupExpose;
+} satisfies HeadingSetupExpose;
 
 defineExpose(exposed);
 </script>
 
 <style scoped>
-/* Headless by design. Typography and truncation CSS remain consumer-owned. */
+/* Headless by design. Heading scale, rhythm, and truncation CSS remain consumer-owned. */
 </style>
