+++ npm/ui/core/src/families/typography/text/text-types.ts
@@ -1,4 +1,4 @@
-import type { PrimitiveElement } from "./primitive.ts";
+import type { PrimitiveElement } from "../../../primitive.ts";
 
 /** Consumer sizing tokens mirrored by {@link Text} through `data-size`. */
 export type TextSize = "xs" | "sm" | "md" | "lg" | "xl";
