+++ npm/ui/core/src/families/typography/text/text.vue
@@ -1,10 +1,9 @@
 <template>
   <component
-    :is="resolvedAs"
+    :is="as"
     ref="element"
     part="root"
-    data-vize-ui="heading"
-    :data-level="levelState"
+    data-vize-ui="text"
     :data-size="sizeState"
     :data-weight="weightState"
     :data-tone="toneState"
@@ -17,60 +16,50 @@
 <script setup lang="ts">
 import { computed, useTemplateRef } from "vue";
 
-import type { PrimitiveAs } from "./primitive.ts";
+import type { PrimitiveAs } from "../../../primitive.ts";
 import type {
-  HeadingElement,
-  HeadingExpose,
-  HeadingLevel,
-  HeadingSize,
-  HeadingSlotState,
-  HeadingTone,
-  HeadingWeight,
-} from "./heading-types.ts";
+  TextElement,
+  TextExpose,
+  TextSlotState,
+  TextSize,
+  TextTone,
+  TextWeight,
+} from "./text-types.ts";
 
 const {
-  as,
-  level = 2,
+  as = "span",
   size = "md",
-  weight = "semibold",
+  weight = "regular",
   tone = "neutral",
   truncate = false,
 } = defineProps<{
   /**
-   * Native element, custom element, or component to render. When omitted,
-   * Heading renders the native `h${level}` element.
+   * Native element, custom element, or component to render.
    *
-   * @default undefined
+   * @default "span"
    */
   readonly as?: PrimitiveAs;
 
   /**
-   * Semantic heading level used for the default native host and `data-level`.
-   *
-   * @default 2
-   */
-  readonly level?: HeadingLevel;
-
-  /**
-   * Consumer visual-size token mirrored to `data-size`; no CSS is emitted.
+   * Consumer text-size token mirrored to `data-size`; no CSS is emitted.
    *
    * @default "md"
    */
-  readonly size?: HeadingSize;
+  readonly size?: TextSize;
 
   /**
    * Consumer font-weight token mirrored to `data-weight`; no CSS is emitted.
    *
-   * @default "semibold"
+   * @default "regular"
    */
-  readonly weight?: HeadingWeight;
+  readonly weight?: TextWeight;
 
   /**
    * Consumer tone token mirrored to `data-tone`; no CSS is emitted.
    *
    * @default "neutral"
    */
-  readonly tone?: HeadingTone;
+  readonly tone?: TextTone;
 
   /**
    * Consumer truncation hook mirrored to `data-truncate`; no CSS is emitted.
@@ -81,31 +70,24 @@ const {
 }>();
 
 defineSlots<{
-  /** Renders heading content with current semantic and typography hook state. */
-  default(props: HeadingSlotState): unknown;
+  /** Renders text content with current typography hook state. */
+  default(props: TextSlotState): unknown;
 }>();
 
-const element = useTemplateRef<HeadingElement>("element");
-const levelState = computed(() => level);
+const element = useTemplateRef<TextElement>("element");
 const sizeState = computed(() => size);
 const weightState = computed(() => weight);
 const toneState = computed(() => tone);
 const truncateState = computed(() => truncate);
-const resolvedAs = computed(() => as ?? `h${levelState.value}`);
-const slotState = computed<HeadingSlotState>(() => ({
-  level: levelState.value,
+const slotState = computed<TextSlotState>(() => ({
   size: sizeState.value,
   tone: toneState.value,
   truncate: truncateState.value,
   weight: weightState.value,
 }));
 
-type HeadingSetupExpose = Omit<
-  HeadingExpose,
-  "element" | "level" | "size" | "tone" | "truncate" | "weight"
-> & {
+type TextSetupExpose = Omit<TextExpose, "element" | "size" | "tone" | "truncate" | "weight"> & {
   readonly element: typeof element;
-  readonly level: typeof levelState;
   readonly size: typeof sizeState;
   readonly tone: typeof toneState;
   readonly truncate: typeof truncateState;
@@ -114,16 +96,15 @@ type HeadingSetupExpose = Omit<
 
 const exposed = {
   element,
-  level: levelState,
   size: sizeState,
   tone: toneState,
   truncate: truncateState,
   weight: weightState,
-} satisfies HeadingSetupExpose;
+} satisfies TextSetupExpose;
 
 defineExpose(exposed);
 </script>
 
 <style scoped>
-/* Headless by design. Heading scale, rhythm, and truncation CSS remain consumer-owned. */
+/* Headless by design. Typography and truncation CSS remain consumer-owned. */
 </style>
