+++ npm/ui/core/src/family-catalog-layout.ts
@@ -41,29 +41,6 @@ export const layoutFamilyCatalog = [
   ...avatarLayoutFamilyCatalog,
   ...iconLayoutFamilyCatalog,
   {
-    canonicalName: "blockquote",
-    title: "Blockquote",
-    packageSubpath: "./blockquote",
-    entryFile: "src/blockquote.ts",
-    sourceFiles: ["src/blockquote.vue", "src/blockquote.ts", "src/blockquote-types.ts"],
-    behaviorContract: "src/blockquote.behavior.md",
-    tests: ["src/blockquote.test.ts", "src/blockquote-ssr.test.ts"],
-    typeTests: ["src/blockquote.types.test-d.ts"],
-    rendererFixture: "BlockquoteConsumer.vue",
-    qualityGates: componentQualityGates,
-    bundleBudget: {
-      exportName: "Blockquote",
-      retainedSignature: 'data-vize-ui":(?:`blockquote`|"blockquote"|\'blockquote\')',
-      maximumJavaScriptGzipBytes: 900,
-      maximumCssGzipBytes: 0,
-    },
-    aliases: ["quote", "pull quote", "quoted text", "quotation block"],
-    upstreamCoverage: ["HTML blockquote element", "HTML cite attribute", "Reka UI Primitive"],
-    dependencies: [],
-    maturity: "stable",
-    owner: catalogOwner,
-  },
-  {
     canonicalName: "card",
     title: "Card",
     packageSubpath: "./card",
@@ -82,29 +59,6 @@ export const layoutFamilyCatalog = [
     },
     aliases: ["surface card", "panel", "content surface", "section shell"],
     upstreamCoverage: ["HTML section", "HTML article", "shadcn/ui Card", "Reka UI Primitive"],
-    dependencies: [],
-    maturity: "stable",
-    owner: catalogOwner,
-  },
-  {
-    canonicalName: "code",
-    title: "Code",
-    packageSubpath: "./code",
-    entryFile: "src/code.ts",
-    sourceFiles: ["src/code.vue", "src/code.ts", "src/code-types.ts"],
-    behaviorContract: "src/code.behavior.md",
-    tests: ["src/code.test.ts", "src/code-ssr.test.ts"],
-    typeTests: ["src/code.types.test-d.ts"],
-    rendererFixture: "CodeConsumer.vue",
-    qualityGates: componentQualityGates,
-    bundleBudget: {
-      exportName: "Code",
-      retainedSignature: 'data-vize-ui":(?:`code`|"code"|\'code\')',
-      maximumJavaScriptGzipBytes: 850,
-      maximumCssGzipBytes: 0,
-    },
-    aliases: ["inline code", "code snippet", "code block", "monospace text"],
-    upstreamCoverage: ["HTML code element", "HTML pre element", "Reka UI Primitive"],
     dependencies: [],
     maturity: "stable",
     owner: catalogOwner,
@@ -270,73 +224,4 @@ export const layoutFamilyCatalog = [
     owner: catalogOwner,
   },
   ...surfaceLayoutFamilyCatalog,
-  {
-    canonicalName: "heading",
-    title: "Heading",
-    packageSubpath: "./heading",
-    entryFile: "src/heading.ts",
-    sourceFiles: ["src/heading.vue", "src/heading.ts", "src/heading-types.ts"],
-    behaviorContract: "src/heading.behavior.md",
-    tests: ["src/heading.test.ts", "src/heading-ssr.test.ts"],
-    typeTests: ["src/heading.types.test-d.ts"],
-    rendererFixture: "HeadingConsumer.vue",
-    qualityGates: componentQualityGates,
-    bundleBudget: {
-      exportName: "Heading",
-      retainedSignature: 'data-vize-ui":(?:`heading`|"heading"|\'heading\')',
-      maximumJavaScriptGzipBytes: 950,
-      maximumCssGzipBytes: 0,
-    },
-    aliases: ["semantic heading", "typography heading", "title", "section heading"],
-    upstreamCoverage: ["HTML heading content", "ARIA document structure", "Reka UI Primitive"],
-    dependencies: [],
-    maturity: "stable",
-    owner: catalogOwner,
-  },
-  {
-    canonicalName: "text",
-    title: "Text",
-    packageSubpath: "./text",
-    entryFile: "src/text.ts",
-    sourceFiles: ["src/text.vue", "src/text.ts", "src/text-types.ts"],
-    behaviorContract: "src/text.behavior.md",
-    tests: ["src/text.test.ts", "src/text-ssr.test.ts"],
-    typeTests: ["src/text.types.test-d.ts"],
-    rendererFixture: "TextConsumer.vue",
-    qualityGates: componentQualityGates,
-    bundleBudget: {
-      exportName: "Text",
-      retainedSignature: 'data-vize-ui":(?:`text`|"text"|\'text\')',
-      maximumJavaScriptGzipBytes: 900,
-      maximumCssGzipBytes: 0,
-    },
-    aliases: ["body text", "typography text", "copy", "inline text", "truncated text"],
-    upstreamCoverage: ["HTML phrasing content", "CSS text-overflow", "Reka UI Primitive"],
-    dependencies: [],
-    maturity: "stable",
-    owner: catalogOwner,
-  },
-  {
-    canonicalName: "kbd",
-    title: "Kbd",
-    packageSubpath: "./kbd",
-    entryFile: "src/kbd.ts",
-    sourceFiles: ["src/kbd.vue", "src/kbd.ts", "src/kbd-types.ts"],
-    behaviorContract: "src/kbd.behavior.md",
-    tests: ["src/kbd.test.ts", "src/kbd-ssr.test.ts"],
-    typeTests: ["src/kbd.types.test-d.ts"],
-    rendererFixture: "KbdConsumer.vue",
-    qualityGates: componentQualityGates,
-    bundleBudget: {
-      exportName: "Kbd",
-      retainedSignature: 'data-vize-ui":(?:`kbd`|"kbd"|\'kbd\')',
-      maximumJavaScriptGzipBytes: 850,
-      maximumCssGzipBytes: 0,
-    },
-    aliases: ["keyboard input", "shortcut", "keycap", "keyboard sequence"],
-    upstreamCoverage: ["HTML kbd element", "CSS inline layout", "Reka UI Primitive"],
-    dependencies: [],
-    maturity: "stable",
-    owner: catalogOwner,
-  },
 ] as const satisfies readonly UiFamilyCatalogEntry[];
