+++ npm/ui/core/src/family-catalog-typography.ts
@@ -1,0 +1,158 @@
+import {
+  catalogOwner,
+  componentQualityGates,
+  type UiFamilyCatalogEntry,
+} from "./family-catalog-types.ts";
+
+export const typographyFamilyCatalog = [
+  {
+    canonicalName: "blockquote",
+    title: "Blockquote",
+    packageSubpath: "./blockquote",
+    entryFile: "src/families/typography/blockquote/blockquote.ts",
+    sourceFiles: [
+      "src/families/typography/blockquote/blockquote.vue",
+      "src/families/typography/blockquote/blockquote.ts",
+      "src/families/typography/blockquote/blockquote-types.ts",
+    ],
+    behaviorContract: "src/families/typography/blockquote/blockquote.behavior.md",
+    tests: [
+      "src/families/typography/blockquote/blockquote.test.ts",
+      "src/families/typography/blockquote/blockquote-ssr.test.ts",
+    ],
+    typeTests: ["src/families/typography/blockquote/blockquote.types.test-d.ts"],
+    rendererFixture: "BlockquoteConsumer.vue",
+    qualityGates: componentQualityGates,
+    bundleBudget: {
+      exportName: "Blockquote",
+      retainedSignature: 'data-vize-ui":(?:`blockquote`|"blockquote"|\'blockquote\')',
+      maximumJavaScriptGzipBytes: 900,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["quote", "pull quote", "quoted text", "quotation block"],
+    upstreamCoverage: ["HTML blockquote element", "HTML cite attribute", "Reka UI Primitive"],
+    dependencies: [],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "code",
+    title: "Code",
+    packageSubpath: "./code",
+    entryFile: "src/families/typography/code/code.ts",
+    sourceFiles: [
+      "src/families/typography/code/code.vue",
+      "src/families/typography/code/code.ts",
+      "src/families/typography/code/code-types.ts",
+    ],
+    behaviorContract: "src/families/typography/code/code.behavior.md",
+    tests: [
+      "src/families/typography/code/code.test.ts",
+      "src/families/typography/code/code-ssr.test.ts",
+    ],
+    typeTests: ["src/families/typography/code/code.types.test-d.ts"],
+    rendererFixture: "CodeConsumer.vue",
+    qualityGates: componentQualityGates,
+    bundleBudget: {
+      exportName: "Code",
+      retainedSignature: 'data-vize-ui":(?:`code`|"code"|\'code\')',
+      maximumJavaScriptGzipBytes: 850,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["inline code", "code snippet", "code block", "monospace text"],
+    upstreamCoverage: ["HTML code element", "HTML pre element", "Reka UI Primitive"],
+    dependencies: [],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "heading",
+    title: "Heading",
+    packageSubpath: "./heading",
+    entryFile: "src/families/typography/heading/heading.ts",
+    sourceFiles: [
+      "src/families/typography/heading/heading.vue",
+      "src/families/typography/heading/heading.ts",
+      "src/families/typography/heading/heading-types.ts",
+    ],
+    behaviorContract: "src/families/typography/heading/heading.behavior.md",
+    tests: [
+      "src/families/typography/heading/heading.test.ts",
+      "src/families/typography/heading/heading-ssr.test.ts",
+    ],
+    typeTests: ["src/families/typography/heading/heading.types.test-d.ts"],
+    rendererFixture: "HeadingConsumer.vue",
+    qualityGates: componentQualityGates,
+    bundleBudget: {
+      exportName: "Heading",
+      retainedSignature: 'data-vize-ui":(?:`heading`|"heading"|\'heading\')',
+      maximumJavaScriptGzipBytes: 950,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["semantic heading", "typography heading", "title", "section heading"],
+    upstreamCoverage: ["HTML heading content", "ARIA document structure", "Reka UI Primitive"],
+    dependencies: [],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "kbd",
+    title: "Kbd",
+    packageSubpath: "./kbd",
+    entryFile: "src/families/typography/kbd/kbd.ts",
+    sourceFiles: [
+      "src/families/typography/kbd/kbd.vue",
+      "src/families/typography/kbd/kbd.ts",
+      "src/families/typography/kbd/kbd-types.ts",
+    ],
+    behaviorContract: "src/families/typography/kbd/kbd.behavior.md",
+    tests: [
+      "src/families/typography/kbd/kbd.test.ts",
+      "src/families/typography/kbd/kbd-ssr.test.ts",
+    ],
+    typeTests: ["src/families/typography/kbd/kbd.types.test-d.ts"],
+    rendererFixture: "KbdConsumer.vue",
+    qualityGates: componentQualityGates,
+    bundleBudget: {
+      exportName: "Kbd",
+      retainedSignature: 'data-vize-ui":(?:`kbd`|"kbd"|\'kbd\')',
+      maximumJavaScriptGzipBytes: 850,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["keyboard input", "shortcut", "keycap", "keyboard sequence"],
+    upstreamCoverage: ["HTML kbd element", "CSS inline layout", "Reka UI Primitive"],
+    dependencies: [],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "text",
+    title: "Text",
+    packageSubpath: "./text",
+    entryFile: "src/families/typography/text/text.ts",
+    sourceFiles: [
+      "src/families/typography/text/text.vue",
+      "src/families/typography/text/text.ts",
+      "src/families/typography/text/text-types.ts",
+    ],
+    behaviorContract: "src/families/typography/text/text.behavior.md",
+    tests: [
+      "src/families/typography/text/text.test.ts",
+      "src/families/typography/text/text-ssr.test.ts",
+    ],
+    typeTests: ["src/families/typography/text/text.types.test-d.ts"],
+    rendererFixture: "TextConsumer.vue",
+    qualityGates: componentQualityGates,
+    bundleBudget: {
+      exportName: "Text",
+      retainedSignature: 'data-vize-ui":(?:`text`|"text"|\'text\')',
+      maximumJavaScriptGzipBytes: 900,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["body text", "typography text", "copy", "inline text", "truncated text"],
+    upstreamCoverage: ["HTML phrasing content", "CSS text-overflow", "Reka UI Primitive"],
+    dependencies: [],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+] as const satisfies readonly UiFamilyCatalogEntry[];
