+++ npm/ui/core/src/sfc-lint.test.ts
@@ -27,7 +27,6 @@ test("discovers every SFC with the opinionated Vize contract", async () => {
       "src/avatar.vue",
       "src/badge.vue",
       "src/block-ui.vue",
-      "src/blockquote.vue",
       "src/breadcrumb-item.vue",
       "src/breadcrumb-link.vue",
       "src/breadcrumb-list.vue",
@@ -36,7 +35,6 @@ test("discovers every SFC with the opinionated Vize contract", async () => {
       "src/card.vue",
       "src/checkbox-control.vue",
       "src/cluster.vue",
-      "src/code.vue",
       "src/collapsible-content.vue",
       "src/collapsible-root.vue",
       "src/collapsible-trigger.vue",
@@ -86,13 +84,16 @@ test("discovers every SFC with the opinionated Vize contract", async () => {
       "src/families/overlays/tooltip/tooltip-root.vue",
       "src/families/overlays/tooltip/tooltip-trigger.vue",
       "src/families/selection/native-select/native-select.vue",
+      "src/families/typography/blockquote/blockquote.vue",
+      "src/families/typography/code/code.vue",
+      "src/families/typography/heading/heading.vue",
+      "src/families/typography/kbd/kbd.vue",
+      "src/families/typography/text/text.vue",
       "src/field-description.vue",
       "src/field-error-message.vue",
       "src/field-label.vue",
       "src/field.vue",
       "src/grid.vue",
-      "src/heading.vue",
-      "src/kbd.vue",
       "src/link-anchor.vue",
       "src/list.vue",
       "src/listbox-item.vue",
@@ -131,7 +132,6 @@ test("discovers every SFC with the opinionated Vize contract", async () => {
       "src/tabs-root.vue",
       "src/tabs-trigger.vue",
       "src/text-input.vue",
-      "src/text.vue",
       "src/textarea-control.vue",
       "src/toggle-button.vue",
       "src/toggle-group-item.vue",
@@ -151,7 +151,6 @@ test("discovers every SFC with the opinionated Vize contract", async () => {
       "avatar.vue",
       "badge.vue",
       "block-ui.vue",
-      "blockquote.vue",
       "breadcrumb-item.vue",
       "breadcrumb-link.vue",
       "breadcrumb-list.vue",
@@ -160,7 +159,6 @@ test("discovers every SFC with the opinionated Vize contract", async () => {
       "card.vue",
       "checkbox-control.vue",
       "cluster.vue",
-      "code.vue",
       "collapsible-content.vue",
       "collapsible-root.vue",
       "collapsible-trigger.vue",
@@ -210,13 +208,16 @@ test("discovers every SFC with the opinionated Vize contract", async () => {
       "families/overlays/tooltip/tooltip-root.vue",
       "families/overlays/tooltip/tooltip-trigger.vue",
       "families/selection/native-select/native-select.vue",
+      "families/typography/blockquote/blockquote.vue",
+      "families/typography/code/code.vue",
+      "families/typography/heading/heading.vue",
+      "families/typography/kbd/kbd.vue",
+      "families/typography/text/text.vue",
       "field-description.vue",
       "field-error-message.vue",
       "field-label.vue",
       "field.vue",
       "grid.vue",
-      "heading.vue",
-      "kbd.vue",
       "link-anchor.vue",
       "list.vue",
       "listbox-item.vue",
@@ -255,7 +256,6 @@ test("discovers every SFC with the opinionated Vize contract", async () => {
       "tabs-root.vue",
       "tabs-trigger.vue",
       "text-input.vue",
-      "text.vue",
       "textarea-control.vue",
       "toggle-button.vue",
       "toggle-group-item.vue",
