+++ npm/ui/core/vite.config.ts
@@ -108,12 +108,12 @@ export default defineConfig({
       banner: "src/families/feedback/banner/banner.ts",
       "block-ui": "src/block-ui.ts",
       callout: "src/families/feedback/callout/callout.ts",
-      blockquote: "src/blockquote.ts",
+      blockquote: "src/families/typography/blockquote/blockquote.ts",
       breadcrumb: "src/breadcrumb.ts",
       tabs: "src/tabs.ts",
       stepper: "src/stepper.ts",
       card: "src/card.ts",
-      code: "src/code.ts",
+      code: "src/families/typography/code/code.ts",
       cluster: "src/cluster.ts",
       container: "src/container.ts",
       grid: "src/grid.ts",
@@ -146,12 +146,12 @@ export default defineConfig({
       skeleton: "src/skeleton.ts",
       meter: "src/meter.ts",
       "native-select": "src/families/selection/native-select/native-select.ts",
-      heading: "src/heading.ts",
-      kbd: "src/kbd.ts",
+      heading: "src/families/typography/heading/heading.ts",
+      kbd: "src/families/typography/kbd/kbd.ts",
       list: "src/list.ts",
       listbox: "src/listbox.ts",
       pagination: "src/pagination.ts",
-      text: "src/text.ts",
+      text: "src/families/typography/text/text.ts",
       textarea: "src/textarea.ts",
       switch: "src/switch.ts",
       checkbox: "src/checkbox.ts",
